Basic voicings refactoring
